# apps/CardHero.NetCoreApp.TypeScript/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Card Hero API",
    "description": "Card Hero OpenAPI document.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/store": {
      "get": {
        "tags": [
          "StoreApi"
        ],
        "operationId": "StoreApi_Get",
        "parameters": [
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "Sort",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/StoreItemModel"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/store/{id}/buy": {
      "post": {
        "tags": [
          "StoreApi"
        ],
        "operationId": "StoreApi_BuyStoreItem",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CardCollectionModel"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorViewModel"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorViewModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string",
            "nullable": true
          },
          "stackTrace": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CardModel": {
        "type": "object",
        "description": "Card.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id.",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Description.",
            "nullable": true
          },
          "upAttack": {
            "type": "integer",
            "description": "Up attack.",
            "format": "int32"
          },
          "rightAttack": {
            "type": "integer",
            "description": "Right attack.",
            "format": "int32"
          },
          "downAttack": {
            "type": "integer",
            "description": "Down attack.",
            "format": "int32"
          },
          "leftAttack": {
            "type": "integer",
            "description": "Left attack.",
            "format": "int32"
          },
          "health": {
            "type": "integer",
            "description": "Health.",
            "format": "int32"
          },
          "attack": {
            "type": "integer",
            "description": "Attack.",
            "format": "int32"
          },
          "defence": {
            "type": "integer",
            "description": "Defence.",
            "format": "int32"
          },
          "totalStats": {
            "type": "integer",
            "description": "Total stats.",
            "format": "int32"
          },
          "isFavourited": {
            "type": "boolean",
            "description": "Is favourited."
          },
          "rarity": {
            "description": "Rarity.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/Rarity"
              }
            ]
          }
        }
      },
      "Rarity": {
        "type": "integer",
        "description": "Rarity.",
        "x-enumNames": [
          "Common",
          "Uncommon",
          "Rare",
          "Epic",
          "Legendary"
        ],
        "enum": [
          1,
          2,
          3,
          4,
          5
        ]
      },
...
      "CardCollectionModel": {
        "type": "object",
        "description": "Card collection.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id.",
            "format": "int32"
          },
          "cardId": {
            "type": "integer",
            "description": "Card id.",
            "format": "int32"
          },
          "userId": {
            "type": "integer",
            "description": "User id.",
            "format": "int32"
          },
          "card": {
            "description": "Card.",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CardModel"
              }
            ]
          }
        }
      },
...
      "StoreItemModel": {
        "type": "object",
        "description": "Store item.",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "description": "Id.",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "description": "Name.",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Description.",
            "nullable": true
          },
          "cost": {
            "type": "integer",
            "description": "Cost.",
            "format": "int32"
          },
          "itemCount": {
            "type": "integer",
            "description": "Item count.",
            "format": "int32"
          },
          "expiry": {
            "type": "string",
            "description": "Expiry.",
            "format": "date-time",
            "nullable": true
          }
        }
      }
    }
  }
}